requirements.txt
asgiref~=3.7.2
Flask==3.0.3
python-dotenv==1.0.1
python-telegram-bot==21.6
redis==6.1.1
uvicorn~=0.23.2
